Add ability to hide maildirs with no unread emails.
This is done by introducing a new option,
`mu4e-maildirs-extension-hide-empty-maildirs` set to `nil` by
default (old behaviour).  When the option is set to `t`, the maildir
is not inserted if it doesn't contain new emails.

This requires to also change one of the existing
`mu4e-maildirs-extension-after-insert-maildir-hook` option value to
ensure the following newline is also conditionally inserted.